Adding the hasObstacleAtPosition to the Tunnel class.

+++ Project/Tunnel.h
@@ -25,7 +25,7 @@ public:
 	Tunnel(float obstacle_probability);
 	virtual ~Tunnel();
 	void draw(int c);
-	bool hasObstacleAtPosition(int position);
+	bool hasObstacleAtPosition(float position);
 private:
 	void pushRing();
 	float offset;
